# services/offer/openapi.yaml
...
openapi: 3.0.3
info:
  title: Offer Service
  version: 0.1.0
  license:
    name: Elastic-Internal
servers:
  - url: /api/v1
paths:
...
  /offers/{id}/status:
    put:
      operationId: updateOfferStatus
      summary: Manually update offer status and record history
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OfferStatusUpdate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OfferStatusUpdated' }
        '400':
          description: Invalid input
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OfferStatusUpdate:
      type: object
      properties:
        status: { type: string }
      required: [status]
    OfferStatusUpdated:
      type: object
      properties:
        status: { type: string, enum: [ok] }
        offerId: { type: string }
        from: { type: string }
        to: { type: string }
      required: [status, offerId, from, to]
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
            traceId: { type: string }
          required: [code, message]